requirements.txt
cupy-cuda12x
numpy
opencv-python
pytest
